fix: change Version u8 to u16
We are abusing es version 2015 as `Version(2015, 0, 0)` in oxc so u8
doesn't fit.

+++ src/queries/node_accurate.rs
@@ -10,9 +10,9 @@ pub(super) fn node_accurate(version_str: &str, opts: &Opts) -> QueryResult {
     }
 
     let mut s = version_str.split('.');
-    let major = s.next().map(|n| n.parse::<u8>().unwrap_or_default());
-    let minor = s.next().map(|n| n.parse::<u8>().unwrap_or_default());
-    let patch = s.next().map(|n| n.parse::<u8>().unwrap_or_default());
+    let major = s.next().map(|n| n.parse::<u16>().unwrap_or_default());
+    let minor = s.next().map(|n| n.parse::<u16>().unwrap_or_default());
+    let patch = s.next().map(|n| n.parse::<u16>().unwrap_or_default());
 
     let distribs = NODE_VERSIONS
         .iter()
